requirements.txt
PyMuPDF>=1.22.0
python-docx>=0.8.11
python-pptx>=0.6.21